--- a/src/execution/vm/bytecode_generator.cpp
+++ b/src/execution/vm/bytecode_generator.cpp
@@ -3505,14 +3505,23 @@ void BytecodeGenerator::VisitMemberExpr(ast::MemberExpr *node) {
   // object is zero, we needn't do anything - we can just reinterpret the object
   // pointer. If the field offset is greater than zero, we generate a LEA.
 
-  LocalVar field_ptr;
-  if (offset == 0) {
-    field_ptr = obj_ptr;
-  } else {
-    field_ptr = GetCurrentFunction()->NewLocal(node->GetType()->PointerTo());
-    GetEmitter()->EmitLea(field_ptr, obj_ptr, offset);
-    field_ptr = field_ptr.ValueOf();
-  }
+  // TODO(Kyle): the optimization described above is ultimately what leads
+  // us to attempt a non-type-system-compliant store when compiling the
+  // resulting bytecode to LLVM IR; eliding the optimization entirely solves
+  // the issue here at the bytecode level, but always incurs the cost of an
+  // additional Lea bytecode as well as an additional store in the IR
+
+//  LocalVar field_ptr;
+//  if (offset == 0) {
+//    field_ptr = obj_ptr;
+//  } else {
+//    field_ptr = GetCurrentFunction()->NewLocal(node->GetType()->PointerTo());
+//    GetEmitter()->EmitLea(field_ptr, obj_ptr, offset);
+//    field_ptr = field_ptr.ValueOf();
+//  }
+  LocalVar field_ptr = GetCurrentFunction()->NewLocal(node->GetType()->PointerTo());
+  GetEmitter()->EmitLea(field_ptr, obj_ptr, offset);
+  field_ptr = field_ptr.ValueOf();
 
   if (GetExecutionResult()->IsLValue()) {
     TERRIER_ASSERT(!GetExecutionResult()->HasDestination(), "L-Values produce their destination");