--- a/src/execution/vm/llvm_engine.cpp
+++ b/src/execution/vm/llvm_engine.cpp
@@ -731,7 +731,8 @@ void LLVMEngine::CompiledModuleBuilder::DefineFunction(const FunctionInfo &func_
         case OperandType::LocalCount: {
           std::vector<LocalVar> locals;
           iter.GetLocalCountOperand(i, &locals);
-          for (const auto local : locals) {
+          // TODO(Kyle): need to make & in order to compile?
+          for (const auto& local : locals) {
             args.push_back(locals_map.GetArgumentById(local));
           }
           break;