Fix for issue [#1099](https://github.com/cmu-db/noisepage/issues/1099).

Parallel static aggregations fail in compiled mode. A simple procedure to reproduce follows:

```
CREATE TABLE test (a INT, b INT);
INSERT INTO test (a , b) VALUES (1, 2);
SELECT SUM(a), SUM(b) FROM test;
```

Executing the final query produces the following error messages:

```
[2020-09-17 18:31:34.843] [execution_logger] [error] ERROR IN MODULE:
Stored value type does not match pointer operand type!
  store %1* %16, %"class.terrier::execution::sql::IntegerSumAggregate"** %5
 %"class.terrier::execution::sql::IntegerSumAggregate"*Stored value type does not match pointer operand type!
  store %1* %22, %"class.terrier::execution::sql::IntegerSumAggregate"** %8
 %"class.terrier::execution::sql::IntegerSumAggregate"*Stored value type does not match pointer operand type!
  store %1* %5, %"class.terrier::execution::sql::IntegerSumAggregate"** %2
 %"class.terrier::execution::sql::IntegerSumAggregate"*Stored value type does not match pointer operand type!
  store %1* %7, %"class.terrier::execution::sql::IntegerSumAggregate"** %3
 %"class.terrier::execution::sql::IntegerSumAggregate"*Stored value type does not match pointer operand type!
  store %1* %31, %"class.terrier::execution::sql::IntegerSumAggregate"** %10
 %"class.terrier::execution::sql::IntegerSumAggregate"*Stored value type does not match pointer operand type!
  store %1* %35, %"class.terrier::execution::sql::IntegerSumAggregate"** %12
 %"class.terrier::execution::sql::IntegerSumAggregate"*
```

The majority of this error message is produced by the `Verifier::visitStoreInst()` function in `Verifier.cpp` within LLVM. As the message suggests, we generate IR with a type mismatch between the stored value and the location (address) at which it is stored. The (simplified) syntax for an LLVM IR `store` looks like: `store <ty> <value> <ty>* <pointer>` and the LLVM type system (enforced in part by `Verifier.cpp`) requires that "the type of the <pointer> operand must be a pointer to the first class type of the <value> operand". 

The issue arises in the `MergeAggregates` function. We produce the following bytecode for `MergeAggregates`:

```
Function 0 <MergeAggregates>:
  Frame size 48 bytes (2 parameters, 6 locals)
    param     queryState:  offset=0       size=8       align=8       type=*struct{*ExecutionContext,struct{IntegerSumAggregate,IntegerSumAggregate}}
    param  pipelineState:  offset=8       size=8       align=8       type=*struct{struct{IntegerSumAggregate,IntegerSumAggregate}}
    local           tmp1:  offset=16      size=8       align=8       type=*struct{IntegerSumAggregate,IntegerSumAggregate}
    local           tmp2:  offset=24      size=8       align=8       type=*struct{IntegerSumAggregate,IntegerSumAggregate}
    local           tmp3:  offset=32      size=8       align=8       type=*IntegerSumAggregate
    local           tmp4:  offset=40      size=8       align=8       type=*IntegerSumAggregate

  0x00000000    Lea                                             local=&tmp1  local=queryState  i32=8
  0x00000010    IntegerSumAggregateMerge                        local=tmp1  local=pipelineState
  0x0000001c    Lea                                             local=&tmp2  local=queryState  i32=8
  0x0000002c    Lea                                             local=&tmp3  local=tmp2  i32=16
  0x0000003c    Lea                                             local=&tmp4  local=pipelineState  i32=16
  0x0000004c    IntegerSumAggregateMerge                        local=tmp3  local=tmp4
  0x00000058    Return  
```

The error occurs at the final `Lea` bytecode: `Lea &tmp4, pipelineState, 16`. Within the LLVM engine, the `Lea` TPL instruction is implemented in terms of a `store` in LLVM IR.

We have that the `pipelineState` type looks like:

```
struct {
    struct {
        IntegerSumAggregate,
        IntegerSumAggregate
    }
}
```

While the type of local `tmp4` is simply `*IntegerSumAggregate`.

To construct the `store` instruction from the bytecode, we perform the following procedure:

- Get the offset from the 2nd (0-indexed) operand to the `Lea` instruction (16)
- Compute the element within `pipelineState` that contains this offset; `pipelineState` is a structure with only a single member (itself a structure) so the result of this computation is the sole member of `pipelineState` with type: `struct {IntegerSumAggregate, IntegerSumAggregate}`
- Compute the index of this element within the top level structure (`pipelineState`); this calculation results in 0 as the index
- Compute a pointer to the element at the index computed above; this results in a pointer with type `*struct {IntegerSumAggregate, IntegerSumAggregate}`
- Create a `store` instruction that stores this computed pointer to `&tmp4`

The resulting `store` IR therefore looks like (in pseudo-IR):

```
store *struct{IntegerSumAggregate, IntegerSumAggregate} val, **IntegerSumAggregate ptr
```

`Verifier::visitStoreInst()` contains an assertion that verifies (essentially) the following type constraint:

```
typeof(*ptr) == typeof(val)
```

We fail this assertion because we attempt to store a pointer to the structure `struct {IntegerSumAggregate, IntegerSumAggregate}` to a location that expects a pointer to a lone `IntegerSumAggregate`.