@turingcompl33t 2021-01-21
Some preliminary results from the benchmarks supported by OLTPBench. About to run through TPC-H benchmarks with the setup suggested by @thepinetree. Results for the benchmarks are averaged over 5 runs. 

**Average Bytecode Count**

| Benchmark | Master | Optimization Removed |
| --- | --- | --- |
| TPC-C | 124 | 141 |
| Smallbank | 58 | 67 |
| TATP | 197 | 227 |

**Benchmarks (Interpreted)**

| Benchmark | Master | Optimization Removed |
| --- | --- | --- |
| TPC-C | 337 tps | 337 tps |
| Smallbank | 595 tps | 590 tps |
| TATP | 442 tps | 426 tps | 

**Benchmarks (Compiled)**

| Benchmark | Master | Optimization Removed |
| --- | --- | --- |
| TPC-C | 337 tps | 327 tps |
| Smallbank | 594 tps | 588 tps |
| TATP | 545 tps | 537 tps |