@turingcompl33t 2021-01-21
Results from running TPC-H benchmarks comparing the current state of `master` and this PR branch. I ran these benchmarks using the "custom" TPC-H setup found in the latest commit on @thepinetree's [branch](https://github.com/thepinetree/noisepage/commit/490280fb361057699319d96e60a06a6ef4c9feed). Scale factor is 1. Each query is executed a minimum of 100 times in each benchmark. Runtimes are averaged over 10 runs; the standard deviation among individual measurements is included following the average runtime (in parentheses). 

Benchmarks run on Ubuntu 20.04 server instances (c5.9xlarge) on AWS with the following specs:

- 18 X 3404.27 MHz CPUs
- L1 Data 32 KiB (x18)
- L1 Instruction 32 KiB (x18)
- L2 Unified 1024 KiB (x18)
- L3 Unified 25344 KiB (x1)

**Interpreted**

| Master | PR | _Diff_ |
| -- | -- | -- |
| 77392ms (759) | 77182ms (570) | -0.27% |

**Compiled**

| Master | PR | _Diff_ |
| -- | -- | -- |
| 56000ms (42) | 55961ms (87) | -0.07% |

In both interpreted and compiled execution modes, we see a very slight decrease in the overall runtime of the benchmark. Obviously it makes no sense that _adding_ additional byetcode to be interpreted / compiled makes the system run faster; I think the takeaway from these results is: the overhead of introducing the additional bytecode (rather, not eliding the `Lea` in circumstances in which it is possible) has such a negligible impact on overall performance that the difference is lost in the noise of the system, despite the fact that Google Benchmark works hard to give us reproducible results and all of the benchmarks were run multiple times independently.

Interested to hear what other's takes may be. I also think it may be worthwhile at this point to mark this "ready for CI" to determine if perfguard uncovers anything different, although I suppose it has also been showing spurious changes in benchmark performance, so it is difficult to arrive at a definitive answer.