@pmenon 2021-01-22
Nice analysis! Just a few points:

1. Can you provide performance + avg. bytecode counts for TPC-H queries individually?
2. Overhead in other parts of the system are most likely hiding any potential overhead incurred by this change - TPC-H SF1 should complete far faster than 56s. Can you hand-code a TPL program that stresses your modification and compare with master?
3. If you want to add more testing, pull the codegen harness from upstream TPL [here](https://github.com/pmenon/tpl/blob/icg/test/include/util/codegen_test_harness.h). It runs codegen tests in all permutations of parallelism and execution modes.