@turingcompl33t 2021-01-30
Bytecode counts for individual TPC-H queries. Need to qualify these results slightly because they are computed based on the query plans that @malin1993ml manually constructed (see [this branch](https://github.com/thepinetree/noisepage/tree/workload-benchmarks) for reference).

| Query Number | Master Bytecode Count | PR Bytecode Count | Percentage Increase |
| --- | --- | --- | --- |
| 1 | 837 | 931 | +11% |
| 4 | 912 | 1036 | +13% |
| 5 | 1790 | 2032 | +13% |
| 6 | 316 | 355 | +12% |
| 7 | 1814 | 2032 | +12% |
| 11 | 1749 | 1988 | +13% |
| 16 | 1145 | 1268 | +10% |
| 18 | 1643 | 1863 | +13% | 
| 19 | 677 | 748 | +10% |

The percentage increase in the number of bytecode instructions is relatively consistent across the individual queries (or at least, more consistent than I anticipated). Furthermore, the increases are also larger in magnitude than I expected, at least based on the previous runtime benchmark results - a ~10% increase in the number of bytecode instructions seems less trivial than what I have been making it out to be. This is likely the type of result that you may have been expecting to see @pmenon, or at least the kind that you'd want to be aware of before anything gets merged.

I am not super familiar with the internals of the interpreter, so I can't say with any authority how worrying a 10% increase in bytecode instructions is in that context. However, I can definitely say that such an increase _is_ worrying in the context of machine code. Each of these additional bytecodes corresponds to an additional `Lea` in the LLVM IR; there is of course the question of whether LLVM might be able to optimize any of these additional instructions away (do we run any optimization passes in non-adapative compiled mode?) so this might not translate directly to a 10% increase in machine instructions, but it is still reason enough to warrant a second look.

Is eliding this optimization prohibitively expensive? Should I revisit the solution? The alternative that comes to mind is somehow manually "collapsing" the inconsistencies at the LLVM type-system level during bytecode generation, but I haven't actually pursued this yet to see how feasible it would be.